FIX: Don’t create posts for trashed topics
Currently there’s a bug when creating digests: if an existing topic
linked to a saved search has been trashed, then we try to create a new
post linked to this topic and the `PostCreator` object raises an error
since it can’t find the provided topic.

This patch addresses this issue simply by adding a join to the query
that is finding the existing topic custom field. This way if the related
topic has been trashed then we’ll create a new topic instead of trying
to reuse the trashed one.